% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fold_change.R
\name{compute_fold_change}
\alias{compute_fold_change}
\title{Compute the expression fold changes for every gene of interest}
\usage{
compute_fold_change(
  data,
  group_var = "sample_id",
  ct_var = "mean_ct",
  ref_nm,
  cal_nm,
  method = "ddct",
  use_geomean = FALSE,
  efficiencies = NULL
)
}
\arguments{
\item{data}{data.frame / tibble. Must contain the columns group_var, "tar_nm" and ct_var.}

\item{group_var}{character vector of length 1. Name of the column to use as grouping variable. Default is "sample_id".}

\item{ct_var}{character vector of length 1. Name of the column containing the Ct values. Default is "mean_id".}

\item{ref_nm}{character vector. Name(s) of the gene(s) to use as reference gene(s). All ref_nm elements must be in the "tar_nm" column of the input data.frame.}

\item{cal_nm}{character vector of length 1. Name of the sample to use as calibrator. Must be in the group_var column of the input data.frame.}

\item{method}{character vector of length 1. Indicates which method to use for computation of gene expression fold changes. Valid values are "ddct" or "Pfaffl".}

\item{use_geomean}{logical vector of length 1. Indicates whether to use geometric mean instead of arithmetic mean for computing the mean of multiple reference genes Ct values. Ignored if method is "Pfaffl". Default is FALSE.}

\item{efficiencies}{data.frame / tibble. Table with gene-specific primer efficiencies. Must contain the columns "tar_nm" and "efficiency", and one entry for each gene in the "tar_nm" column of the input data.frame. Ignored if method is "ddct". If missing and method is "Pfaffl", efficiency is set to 2 for every gene.}
}
\value{
A data.frame of variable length: 10 (ddct method), 12 (Pfaffl method and multiple ref_nm), or 14 (Pfaffl method and single ref_nm). Gene expression fold changes are stored in the "fold_change" column.
}
\description{
compute_fold_change() calculates gene expression fold changes, provided a data.frame, a grouping variable, a calibrator sample and one or more reference gene(s).
}
