% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean_ct.R
\name{compute_mean_ct}
\alias{compute_mean_ct}
\title{Compute the mean Ct value for each level of a grouping variable}
\usage{
compute_mean_ct(data, group_var = "sample_id", use_geomean = FALSE)
}
\arguments{
\item{data}{data.frame / tibble. Must contain the columns group_var, "tar_nm" and "ct".}

\item{group_var}{character vector of length 1. Name of the column to use as grouping variable. Default is "sample_id".}

\item{use_geomean}{logical vector of length 1. Indicates whether to compute geometric mean instead of arithmetic mean. Default is FALSE.}
}
\value{
A data.frame with 7 columns: group_var, "tar_nm", "raw_data", "n_rep", "n_filt_rep", "mean_ct" and "sd_ct".
}
\description{
compute_mean_ct() calculates the arithmetic or geometric mean of Ct values, provided a data.frame and a grouping variable.
}
