% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_amp.R
\name{import_amp}
\alias{import_amp}
\title{Import amplification data from QuantStudio & StepOne qPCR software}
\usage{
import_amp(file, from = "QuantStudio")
}
\arguments{
\item{file}{character vector of length 1. Path to the .xls file containing qPCR results.}

\item{from}{character vector of length 1. Indicates which qPCR software the input file originates from. Valid values are "QuantStudio" or "StepOne".}
}
\value{
A tibble with 4 columns: "well_position", "sample_id", "tar_nm", "cycle", "rn", and "delta_rn".
}
\description{
import_amp() tidies the "Amplification Data" sheet of an .xls file produced by QuantStudio Design & Analysis Software v1.5.2 or StepOne Software v2.3.
}
