% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_melt.R
\name{import_melt}
\alias{import_melt}
\title{Import melt curve data from QuantStudio qPCR software}
\usage{
import_melt(file, from = "QuantStudio")
}
\arguments{
\item{file}{character vector of length 1. Path to the .xls file containing the qPCR results to import.}

\item{from}{character vector of length 1. Indicates which qPCR software the input file originates from. Currently only supports "QuantStudio".}
}
\value{
A tibble with 7 columns: "well_position", "sample_id", "tar_nm", "reading", "temperature", "fluorescence", "derivative".
}
\description{
import_melt() tidies the "Melt Curve Raw Data" sheet of an .xls file produced by QuantStudio Design & Analysis Software v1.5.2.
}
